fix(logger): handle errors
