+++ lib/logger/utils.ts
@@ -1,6 +1,8 @@
 import { Stream } from 'stream';
 import bunyan from 'bunyan';
 import fs from 'fs-extra';
+import { RequestError } from 'got';
+import { clone } from '../util/clone';
 import { redactedFields, sanitize } from '../util/sanitize';
 
 export interface BunyanRecord extends Record<string, any> {
@@ -49,6 +51,47 @@ const contentFields = [
   'yarnLockParsed',
 ];
 
+export default function prepareError(err: Error): Record<string, unknown> {
+  const response: Record<string, unknown> = {
+    ...err,
+  };
+
+  // Can maybe removed?
+  if (!response.message && err.message) {
+    response.message = err.message;
+  }
+
+  // Can maybe removed?
+  if (!response.stack && err.stack) {
+    response.stack = err.stack;
+  }
+
+  // handle got error
+  if (err instanceof RequestError) {
+    const options: Record<string, unknown> = {
+      headers: clone(err.options.headers),
+      url: err.options.url?.toString(),
+    };
+    response.options = options;
+
+    for (const k of ['username', 'password', 'method', 'http2']) {
+      options[k] = err.options[k];
+    }
+
+    if (err.response) {
+      response.response = {
+        statusCode: err.response?.statusCode,
+        statusMessage: err.response?.statusMessage,
+        body: clone(err.response.body),
+        headers: clone(err.response.headers),
+        httpVersion: err.response.httpVersion,
+      };
+    }
+  }
+
+  return response;
+}
+
 export function sanitizeValue(value: unknown, seen = new WeakMap()): any {
   if (Array.isArray(value)) {
     const length = value.length;
@@ -65,6 +108,11 @@ export function sanitizeValue(value: unknown, seen = new WeakMap()): any {
 
   if (value instanceof Buffer) {
     return '[content]';
+  }
+
+  if (value instanceof Error) {
+    // eslint-disable-next-line no-param-reassign
+    value = prepareError(value);
   }
 
   const valueType = typeof value;
